test(oracle): consolidate E2E coverage
