--- gravity_e2e/cluster_test_cases/binance_price_feed/genesis.toml
@@ -1,74 +1,0 @@
-# Gravity Genesis Configuration - Binance index-kline price feed integration test
-
-[dependencies.genesis_contracts]
-path = "../../../../gravity_chain_core_contracts"
-ref = "codex/sports-score-oracle-poc"
-
-[[genesis_validators]]
-id = "node1"
-address = "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266"
-host = "127.0.0.1"
-validator_port = 45382
-vfn_port = 45392
-stake_amount = "2000000000000000000"
-voting_power = "2000000000000000000"
-consensus_pop = "0x000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000"
-
-[genesis]
-chain_id = 1337
-# Keep the suite fast: task/callback changes are picked up when the JWK
-# consensus epoch manager rebuilds providers for the next epoch.
-epoch_interval_micros = 30000000
-major_version = 1
-consensus_config = "0x0301010a00000000000000280000000000000001010000000a000000000000000100010200000000000000000020000000000000"
-execution_config = "0x00"
-initial_locked_until_micros = 1798848000000000
-
-governance_owner = "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266"
-
-[genesis.faucet]
-address = "0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266"
-balance = "0x2000000000000000000000000000000000000000000000000000000000000000"
-
-[genesis.validator_config]
-minimum_bond = "1000000000000000000"
-maximum_bond = "1000000000000000000000000"
-unbonding_delay_micros = 604800000000
-allow_validator_set_change = true
-voting_power_increase_limit_pct = 20
-max_validator_set_size = "100"
-auto_evict_enabled = false
-auto_evict_threshold_pct = 0
-
-[genesis.staking_config]
-minimum_stake = "1000000000000000000"
-lockup_duration_micros = 86400000000
-unbonding_delay_micros = 86400000000
-
-[genesis.governance_config]
-min_voting_threshold = "1000000000000000000"
-required_proposer_stake = "1000000000000000000"
-voting_duration_micros = 5000000
-
-[genesis.randomness_config]
-variant = 1
-secrecy_threshold = 9223372036854775808
-reconstruction_threshold = 12297829382473033728
-fast_path_secrecy_threshold = 12297829382473033728
-
-[genesis.oracle_config]
-source_types = [1, 3]
-callbacks = [
-  "0x00000000000000000000000000000001625F4001",
-  "0x0000000000000000000000000000000000000000",
-]
-
-[genesis.jwk_config]
-issuers = ["0x68747470733a2f2f6163636f756e74732e676f6f676c652e636f6d"]
-
-[[genesis.jwk_config.jwks]]
-kid = "f5f4c0ae6e6090a65ab0a694d6ba6f19d5d0b4e6"
-kty = "RSA"
-alg = "RS256"
-e = "AQAB"
-n = "2K7epoJWl_aBoYGpXmDBBiEnwQ0QdVRU1gsbGXNrEbrZEQdY5KjH5P5gZMq3d3KvT1j5KsD2tF_9jFMDLqV4VWDNJRLgSNJxhJuO_oLO2BXUSL9a7fLHxnZCUfJvT2K-O8AXjT3_ZM8UuL8d4jBn_fZLzdEI4MHrZLVSaHDvvKqL_mExQo6cFD-qyLZ-T6aHv2x8R7L_3X7E1nGMjKVVZMveQ_HMeXvnGxKf5yfEP0hIQlC_kFm4L_1kV1S0UPmMptZL2qI4VnXqmqI6TZJyE-3VXHgNn1Z1O_9QZlPC0fF0spLHf2S3nNqI0v3k2E7q3DkqxVf5xvn7q_X-gPqzVE9Jw"
