--- gravity_e2e/cluster_test_cases/binance_price_feed/relayer_config.json
@@ -1,6 +1,0 @@
-{
-  "uri_mappings": {
-    "gravity://3/1001/price_feed?provider=binance_index_kline_v1&pair=NVDAUSDT&interval=1m&bucketStartMs=1783252500000&decimals=8&graceMs=120000": "http://127.0.0.1:18547",
-    "gravity://3/1002/price_feed?provider=binance_index_kline_v1&pair=TSLAUSDT&interval=1m&bucketStartMs=1783252500000&decimals=8&graceMs=120000": "http://127.0.0.1:18547"
-  }
-}
